Fix std::pair constructor size_t to int conversion in map_generator.cpp
The std::pair<int, int> constructor was implicitly converting size_t values
to int when using brace initialization. Added explicit static_cast<int>()
to avoid MSVC warning C4267 (conversion from size_t to int).

This should fix the remaining Windows CI build failure.

+++ src/map_generator.cpp
@@ -891,7 +891,7 @@ std::vector<std::pair<int, int>> MapGenerator::getSequentialConnections(const st
     std::vector<std::pair<int, int>> connections;
     
     for (size_t i = 1; i < rooms.size(); i++) {
-        connections.push_back({i - 1, i});
+        connections.push_back({static_cast<int>(i - 1), static_cast<int>(i)});
     }
     
     return connections;
